+++ app/src/main/java/com/concordium/wallet/data/model/Token.kt
@@ -1,7 +1,7 @@
 package com.concordium.wallet.data.model
 
 import java.io.Serializable
-import java.math.BigDecimal
+import java.math.BigInteger
 
 // The class must have default value for each field,
 // otherwise Gson fails to use defaults and sets not-nullable fields to null.
@@ -14,8 +14,8 @@ data class Token (
     var contractIndex: String = "",
     var subIndex: String = "",
     var isCCDToken: Boolean = false,
-    var totalBalance: BigDecimal = BigDecimal.ZERO,
-    var atDisposal: BigDecimal = BigDecimal.ZERO,
+    var totalBalance: BigInteger = BigInteger.ZERO,
+    var atDisposal: BigInteger = BigInteger.ZERO,
     var contractName: String = "",
     var symbol: String = ""
 ): Serializable
