+++ app/src/main/java/com/concordium/wallet/data/model/TransferSubmissionStatus.kt
@@ -1,14 +1,14 @@
 package com.concordium.wallet.data.model
 
-import java.math.BigDecimal
+import java.math.BigInteger
 
 data class TransferSubmissionStatus(
     val status: TransactionStatus,
     val outcome: TransactionOutcome?,
-    val amount: BigDecimal?,
+    val amount: BigInteger?,
     val sender: String?,
     val to: String?,
-    val cost: BigDecimal?,
+    val cost: BigInteger?,
     val transactionHash: String?,
     val blockHashes: List<String>?,
     val rejectReason: String?,
