+++ app/src/main/java/com/concordium/wallet/ui/account/accountsoverview/AccountsOverviewViewModel.kt
@@ -24,7 +24,7 @@ import com.concordium.wallet.ui.account.common.accountupdater.AccountUpdater
 import com.concordium.wallet.ui.account.common.accountupdater.TotalBalancesData
 import com.concordium.wallet.util.Log
 import kotlinx.coroutines.launch
-import java.math.BigDecimal
+import java.math.BigInteger
 
 class AccountsOverviewViewModel(application: Application) : AndroidViewModel(application) {
     private val _waitingLiveData = MutableLiveData<Boolean>()
@@ -166,7 +166,7 @@ class AccountsOverviewViewModel(application: Application) : AndroidViewModel(app
             if (identityCount == 0) {
                 _stateLiveData.value = State.NO_IDENTITIES
                 // Set balance, because we know it will be 0
-                _totalBalanceLiveData.value = TotalBalancesData(BigDecimal.ZERO, BigDecimal.ZERO, BigDecimal.ZERO, false)
+                _totalBalanceLiveData.value = TotalBalancesData(BigInteger.ZERO, BigInteger.ZERO, BigInteger.ZERO, false)
                 if(notifyWaitingLiveData){
                     _waitingLiveData.value = false
                 }
@@ -175,7 +175,7 @@ class AccountsOverviewViewModel(application: Application) : AndroidViewModel(app
                 if (accountCount == 0) {
                     _stateLiveData.value = State.NO_ACCOUNTS
                     // Set balance, because we know it will be 0
-                    _totalBalanceLiveData.value = TotalBalancesData(BigDecimal.ZERO, BigDecimal.ZERO, BigDecimal.ZERO, false)
+                    _totalBalanceLiveData.value = TotalBalancesData(BigInteger.ZERO, BigInteger.ZERO, BigInteger.ZERO, false)
                     if(notifyWaitingLiveData){
                         _waitingLiveData.value = false
                     }
